Apply new typedefs.
git-svn-id: http://svn.osdn.net/svnroot/deeangband@7418 89dd94fa-70ca-40ae-8747-2094edff379b

+++ source/trunk/src/z-term.h
@@ -187,14 +187,14 @@ struct term
 	u16b key_xtra;
 	u16b key_size;
 
-	byte wid;
-	byte hgt;
+	int wid;
+	int hgt;
 
-	byte y1;
-	byte y2;
+	int y1;
+	int y2;
 
-	byte *x1;
-	byte *x2;
+	int *x1;
+	int *x2;
 
 	term_win *old;
 	term_win *scr;
